2nd pass on debian-setup scripts

+++ debian-setup.sh
@@ -9,44 +9,27 @@
 
 # Steps
 #
-# - aquire settings (custom kernel, user names, SSH keys, hostname, networking, resolvers, NTP servers)
-# - set up DNS (PTR, A, MX)
+# - View Network Graph v4/v6 http://bgp.he.net/
+# - Aquire settings (custom kernel, user names, SSH keys, hostname, networking, resolvers, NTP servers)
+# - Set up DNS (PTR, A, MX)
+# - Make up /root/server.yml from /server.yml and from /debian-setup/providers/*.yml
 
 export IMAGE_ARCH="amd64"
 export IMAGE_MACHINE="x86_64"
 export IMAGE_ID="Debian"
 export IMAGE_CODENAME="jessie"
 
+export SETUP_PACKAGES="lsb-release ca-certificates wget debian-archive-keyring apt apt-utils"
+#:ubuntu
+#export SETUP_PACKAGES="lsb-release ca-certificates wget ubuntu-keyring apt apt-utils"
 export SETUP_APTSOURCES_URL_PREFIX="https://github.com/szepeviktor/debian-server-tools/raw/master/package/apt-sources"
 export SETUP_APTSOURCESLIST_URL="${SETUP_APTSOURCES_URL_PREFIX}/${IMAGE_CODENAME}-azure.list"
 #export SETUP_SOURCESLIST_URL="${SETUP_APTSOURCES_URL_PREFIX}/${IMAGE_CODENAME}-cloudfront.list"
 #export SETUP_SOURCESLIST_URL="${SETUP_APTSOURCES_URL_PREFIX}/${IMAGE_CODENAME}-hu.list"
-export SETUP_PACKAGES="lsb-release ca-certificates wget debian-archive-keyring apt apt-utils"
+
 export SETUP_SHYAML_URL="https://github.com/0k/shyaml/raw/master/shyaml"
-#:ubuntu
-#export SETUP_PACKAGES="lsb-release ca-certificates wget ubuntu-keyring apt apt-utils"
 
 export WITHOUT_SYSTEMD="yes"
-
-Error() {
-    echo "ERROR: $(tput bold;tput setaf 7;tput setab 1)${*}$(tput sgr0)" 1>&2
-}
-
-# Download architecture-independent packages
-Getpkg() {
-    local P="$1"
-    local R="${2-sid}"
-    local WEB="https://packages.debian.org/${R}/all/${P}/download"
-    local URL
-
-    URL="$(wget -q -O- "$WEB" | grep -o '[^"]\+ftp.de.debian.org/debian[^"]\+\.deb')"
-
-    [ -z "$URL" ] && return 1
-
-    wget -q -O "${P}.deb" "$URL"
-    dpkg -i "${P}.deb"
-    echo "Ret=$?"
-}
 
 Data() {
     shyaml "$@" < /root/server.yml
@@ -86,11 +69,15 @@ apt-get install -y aptitude
 ./debian-image-normalize.sh
 
 # Remove wheezy packages
-apt-get purge -y libboost-iostreams1.49.0 libdb5.1 libgcrypt11 libgnutls26 \
-    libprocps0 libtasn1-3 libudev0 python2.6 python2.6-minimal
+if [ "$(dpkg-query --showformat="\${Status}" --show "libgnutls26" 2> /dev/null)" == "install ok installed" ]; then
+    apt-get purge -y libboost-iostreams1.49.0 libdb5.1 libgcrypt11 libgnutls26 \
+        libprocps0 libtasn1-3 libudev0 python2.6 python2.6-minimal
+fi
+# Remove ClamAV data
+rm -rf /var/lib/clamav /var/log/clamav || true
 
 # Packages used during setup
-apt-get install -y ssh sudo apt-transport-https python-yaml
+apt-get install -y ssh sudo apt-transport-https virt-what python-yaml
 # Install SHYAML
 wget -nv -O /usr/local/bin/shyaml "$SETUP_SHYAML_URL"
 chmod +x /usr/local/bin/shyaml
@@ -103,11 +90,6 @@ done
 eval "$(grep -h -A 5 "^deb " /etc/apt/sources.list.d/*.list | grep "^#K: " | cut -d " " -f 2-)"
 # Get package lists
 apt-get update -qq
-
-# Provider packages
-if [ -n "$(Data get-value provider-package)" ]; then
-    apt-get install -y $(Data get-values provider-package)
-fi
 
 # Virtualization environment
 debian-setup/virt-what
